pyproject.toml
[project]
name = "chatgpt-dmenu"
version = "0.1.0"
description = "A dmenu wrapper for ChatGPT with context-based prompts"
authors = [{ name = "André Gonzalez", email = "lopescg@gmail.com" }]
readme = "README.md"
license = "MIT"
requires-python = ">=3.8"
dependencies = [
    "openai>=1.0.0",
    "pyyaml",
]

[project.scripts]
chatgpt-dmenu = "chatgpt_dmenu.main:main"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"
fix = true

[tool.ruff.lint]
select = ["ALL"]
ignore = [
  "ANN101",  # Missing type annotation for self
  "ANN102",  # Missing type annotation for cls
  "D",       # pydocstyle docstring checks (optional)
  "COM812",  # Conflicts with black
  "S603",    # For this project it makes sense igonring this, all input is validated
  "TRY301",  # Not necessary for this simple project
]

[tool.ruff.format]
# Let black handle string formatting
quote-style = "double"

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
ignore_missing_imports = true
strict_optional = true

[tool.isort]
profile = "black"
line_length = 88

[project.optional-dependencies]
dev = [
  "black",
  "ruff",
  "mypy",
  "isort",
  "types-PyYAML",
]
